# Binpack continuations stack to improve memory density