@vasilmkd 2021-08-20
I'm curious, why not an `Array[Long]`?