@vasilmkd 2021-08-20
Should we try inlining the methods in `object ByteStack` into `IOFiber`? 😆