@viktorklang 2021-08-21
@vasilmkd

>I'm curious, why not an Array[Long]?

My thinking was that it would be unnecessary to store the current stack count in a Long, and if stored as an Int shifted into the Long, it'd be more operations to read and update it, without providing any better binpacking or array access checks.